--- src/qcm/Choix.java
@@ -1,23 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package qcm;
-
-/**
- *
- * @author aurélien
- */
-public class Choix {
-    String nom;
-    boolean juste;
-    
-    
-    Choix(String nom, boolean juste){
-        this.nom = nom;
-        this.juste = juste;
-    }
-    
-    
-}
