+++ src/qcm/Controleur/Controleur_Connexion.java
@@ -5,7 +5,6 @@
  */
 package qcm.Controleur;
 
-import QCM.Connexion;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
