+++ src/qcm/Modèle/Etudiant.java
@@ -3,12 +3,12 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package qcm;
+package qcm.Modèle;
 
 /**
  *
  * @author aurélien
  */
-public class Note {
-    
+public class Etudiant {
+    String a;
 }
