+++ src/qcm/Vue/Vue_Principale.java
@@ -21,7 +21,7 @@ import javax.swing.JMenu;
 import javax.swing.JMenuBar;
 import javax.swing.JMenuItem;
 import javax.swing.JPanel;
-import qcm.Images;
+import qcm.Modèle.Images;
 
 /**
  *
